web/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg[binary]>=3.0.0
python-dotenv==1.0.0
python-chess>=1.0.0
